backend/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
pyrebase4
pymongo[srv]
motor
pydantic